+++ Marketplace/Marketplace/celery.py
@@ -1,6 +1,5 @@
 import os
 from celery import Celery
-
 # set the default Django settings module for the 'celery' program.
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'Marketplace.settings')
