+++ Marketplace/main/tasks.py
@@ -1,6 +1,7 @@
-from celery import app
-
+from Marketplace.celery import app
+from .logic import send_new_good
+from .models import Product
 
 @app.task
-def add(x, y):
-    return x + y
+def send_celery_mail(Product):
+    send_new_good(Product)
