+++ Marketplace/main/urls.py
@@ -1,6 +1,5 @@
 from django.urls import path
 from . import views
-from .views import EditUserView
 from .views import *
 
 urlpatterns = [
